requirements.txt
streamlit>=1.39.0
faiss-cpu==1.9.0.post1
transformers>=4.44.0
sentence-transformers>=3.0.0
pydantic>=2.9.0
python-dotenv>=1.0.1